feat: remove all explanatory texts (for now at least)
I want to focus on the functions and design of the page instead of writing
long and "deep" messages about how much we do, how great we are and how
big plans we have while there is barely "we" and for the last 2 years
nothing has been shipped (of real product value).
I decided to do this commit instead of implementing "About us" page.