fix: update the width of File component from 700 to 800
In Figma, File Component was defined as having width 700px,
and in frontend it was correctly implemented, but...
that Figma design was incorrect. Files Component (in Figma)
which contained File Instances had all of them set to width
800 (override of 700). So during design of File and Files
Components, File was first assumed to be 700, but then it
turned out that 800 is better and width 800 was set in all
File Instances in Files Component designs but designer (me)
forgot then to also update File Component width. Figma is now
updated and this commit updates the File component on frontned.