fix: prevent running same effects twice in `@/lib` hooks
I'm not sure of why this is happening but I guess React triggers render
twice on purpose, whenever someone calls a setState function.
Therefore, if someone has such component or hook:

  // ...
  const [state, setState] = React.useState(42)

  console.log('ran!')

  React.useEffect(() => {
    console.log(`ran effect with state = ${state}!`)
  }, [state])

  function onSomeEvent() {
    setState(420)
  }
  // ...

then triggering `onSomeEffect()` will result in such console output:

  ran!
  ran effect with state = 420!
  ran!
  ran effect with state = 420!

As you see the component/hook runs twice. Again, I assume it's done
on purpose by React library to force developers to handle such case
as it may occur sporadically in normal legit user interactions with
the app. I assume this but may be wrong. Heard it some time ago somewhere.

Anyway, to prevent that this commit introduces a `useEffectWithoutReruns()`
which basically does below (extening above component/hook example):

  // ...
  const [state, setState] = React.useState(42)
  const lastTrigger = React.useRef(null)

  console.log('ran!')

  React.useEffect(() => {
    // below 2 lines prevent reruns of the same effect
    if (deepEquals(lastTrigger.current, [state])) return;
    lastTrigger.current = [state];

    console.log(`ran effect with state = ${state}!`)
  }, [state])

  function onSomeEvent() {
    setState(420)
  }
  // ...

Ok, that is it. This commit message explains the reasons behind
`useEffectWithoutReruns()` just in case those are incorrect and it comes
up in the future, and someone tries to figure out whether this incorrect
reason is actually a reason for that function, or if there were some
other motives. There were not, literally only this double render is
the reason behind `useEffectWithoutReruns()`.
Also, if in future the behaviour or React changes this commit will
nicely explain the origins of `useEffectWithoutReruns()` and how things
worked "back then". Cheers!