fix: downgrade NextJS 14 to 13 (temporary!)
This is a temporary solution to support running this project on
Raspberry Pi 2 machine (armv7l). There are other ways to make it
work that do not require a downgrade but that was the simplest way
for me at the time with the knowledge, time and focus I had so I
choose downgrade. This change should be reversed in future.
We should be always at latest stable versions of packages.
This is for free-beta deployment, before shipping first version of
production deployment downgrade will be reversed and replaced with
some other solution also for free-beta so that production and free-beta
will use the same code while having everything at latest (or close to
it) versions.